Cleaned up code--READABLE!!1!  Added some polish.
GUI is fleshed out, needs polish.  Overall flow of game fleshed out.
GameOver screen has been implemented but needs to be modified to fit the
overall style of the game.  I'm not too happy with the background of the
main game state, and I am still thinking about what to do about it.